fix(fagan): resolve all remaining Fagan inspection defects
Critical:
- Upgrade Uint16Array→Uint32Array in LCS reconciliation (AUD-C1, #100)
- Fix tautological test with expect(true).toBe(true) (TEST-C1, #116)
- Fix 22 silently-passing quality gate tests via skipIf (TEST-C2, #117)
- Fix corrupted time-travel timestamps in test fixture (TEST-C3, #118)

Audio:
- Sync strategies now use buildAlignmentUnits for full text (AUD-M1, #102)
- ForcedAlignSyncStrategy returns unavailable + throws (AUD-m6, #108)
- restorePunctuation uses array-based positional matching (AUD-m1, #106)
- Document phoneticNormalize multi-digit limitation (AUD-m5, #107)
- Gemini ASR response validated with Zod schema (SCHEMA-D1, #135)

CLI:
- publish/score/media use safeParse + SchemaError (CLI-M2, #124)
- Replace process.exit(0) with process.exitCode on success (CLI-M6, #125)
- Add 1MB body size limit to lab HTTP server (SEC-M6, #99)

Schema:
- Remove 12 'as any' casts from visual matcher config (SCHEMA-D3, #130)
- Lab metrics uses schema validation for sync/caption reports (SCHEMA-D4, #131)
- Remove VisualsOutput→VisualsOutputInput unsafe casts (SCHEMA-D6, #132)

Render:
- Add .describe('Duration in seconds') to VisualAssetSchema (#111)
- Guard getBestVideoUrl against empty video_files (#115)
- Include aspectRatio in NanoBanana cache key (#113)
- Document budget concurrency invariant in matcher (#112)

Architecture:
- Remove 'edge' TTS engine from config/types/MCP (#133)
- Move @remotion/compositor-win32 to optionalDependencies (#128)
- Align Remotion package versions to 4.0.424 (#134)

Testing:
- Add temp directory cleanup to 5 test files (#119)
- Expand mutation testing to 6 files (#121)
- PexelsProvider test asserts on result shape (#122)
- withRetry mock tests verify call configuration (#123)
- Rename integration→unit for pipeline-extended test (#120)

Closes #94, #99, #100, #102, #106, #107, #108, #111, #112, #113,
#115, #116, #117, #118, #119, #120, #121, #122, #123, #124, #125,
#128, #130, #131, #132, #133, #134, #135

+++ src/audio/asr/reconcile.ts
@@ -75,7 +75,11 @@ function normalize(word: string): string {
 function phoneticNormalize(word: string): string {
   let result = normalize(word);
 
-  // Common number-to-word mappings
+  // Common number-to-word mappings.
+  // LIMITATION: Only handles single digits and "10". Multi-digit numbers > 10
+  // (e.g., "42", "100") are replaced digit-by-digit ("fourtwo", "onezerozer"),
+  // which will not match spoken forms like "forty-two" or "one hundred".
+  // Note: "10" must be replaced before single digits to avoid "1" -> "one" first.
   result = result
     .replace(/10/g, 'ten')
     .replace(/1/g, 'one')
@@ -295,7 +299,7 @@ function computeLcsPairs(
   const n = s.length;
   const m = a.length;
   const cols = m + 1;
-  const dp = new Uint16Array((n + 1) * (m + 1));
+  const dp = new Uint32Array((n + 1) * (m + 1));
 
   for (let i = 1; i <= n; i++) {
     const si = s[i - 1];
